functions-metadatos/requirements.txt
firebase_functions~=0.1.0
functions-framework
requests
Pillow